spam_email_detection
We’ve all been the recipient of spam emails before. Spam mail, or junk mail, is a type of email
that is sent to a massive number of users at one time, frequently containing cryptic
messages, scams, or most dangerously, phishing content.



In this Project jupyter notebook is used to build an email spam detector. 
Then, use machine learning to
train the spam detector to recognize and classify emails into spam and non-spam